started integration of the new element-wise arithmetic engine into core
