+++ modules/core/include/opencv2/core/mat.hpp
@@ -54,6 +54,7 @@
 #include "opencv2/core/bufferpool.hpp"
 
 #include <array>
+#include <functional>
 #include <type_traits>
 
 namespace cv
@@ -3856,6 +3857,66 @@ protected:
     size_t idx;
 };
 
+
+/////////////////////////////////// BroadcastOp //////////////////////////////////////
+
+/** @brief Op-agnostic driver for a broadcasting element-wise traversal.
+
+BroadcastOp takes a flat list of operand Mats (it does NOT distinguish inputs from outputs), computes
+the numpy-broadcast iteration space over all of them (channels = innermost dim), partitions it into
+tasks, runs them with parallel_for_, and for each tile hands the per-operand slices to a `body`
+callback. Everything semantic - which array is the output, which kernels run, temp buffers - lives in
+`body`. For a cv::Mat the innermost axis is always contiguous, so after dimension collapse every
+operand's innermost step is in {0,1} (1 = contiguous, 0 = broadcast-scalar) - there is no gather case.
+*/
+struct BroadcastOp
+{
+    //! One operand's slice for the current tile: base pointer + steps in ELEMENTS. stepx in {0,1}
+    //! (1 = contiguous along width, 0 = broadcast-scalar); stepy = step between the `height` rows
+    //! (0 = broadcast). ptr is non-const so the body can write the operand(s) it treats as outputs.
+    struct Slice
+    {
+        void*  ptr   = nullptr;
+        size_t stepy = 0;
+        size_t stepx = 0;
+    };
+
+    //! One 2D tile handed to the body. slices[k] corresponds to arrays[k] (same order); the body reads
+    //! width/height and the per-operand slices and owns all interpretation.
+    struct Tile
+    {
+        int width  = 0;                //!< innermost tile extent (elements)
+        int height = 0;                //!< 2nd-innermost extent (1 unless a 2D tile is handed out)
+        int narrays = 0;
+        const Slice* slices = nullptr; //!< [narrays], valid for the duration of the body call
+    };
+
+    /** @brief Drive a broadcasting element-wise traversal.
+    @param arrays   pointers to the operand Mats (inputs AND outputs, undistinguished); the iteration
+                    space is the numpy-broadcast of all their shapes (channels innermost). Headers must
+                    stay alive for the call - no Mat copies are made.
+    @param narrays  number of operands.
+    @param body     invoked once per tile with that tile's per-operand slices; runs the prepared program.
+                    Per-thread scratch is just locals in the body (declared per call => thread-safe).
+    @param expandChannels  true => channels are an explicit innermost iteration dim, so the body always
+                    sees single-channel data (1<->N channel broadcast handled geometrically). false =>
+                    channels stay folded into the element (esz = full elemSize); the body handles them.
+    @param nstripes parallel_for_ work hint; 0 => derive from the shapes (assuming ~100 cycles/element).
+    */
+    CV_EXPORTS static void run(const Mat* const* arrays, int narrays,
+                               const std::function<void(const Tile&)>& body,
+                               bool expandChannels = false,
+                               double nstripes = 0.);
+};
+
+//! Free-function shorthand for BroadcastOp::run (see BroadcastOp).
+inline void broadcastOp(const Mat* const* arrays, int narrays,
+                        const std::function<void(const BroadcastOp::Tile&)>& body,
+                        bool expandChannels = false,
+                        double nstripes = 0.)
+{
+    BroadcastOp::run(arrays, narrays, body, expandChannels, nstripes);
+}
 
 
 ///////////////////////////////// Matrix Expressions /////////////////////////////////
